+++ core-api/src/main/java/jp/co/soramitsu/core/model/RuntimeConfiguration.kt
@@ -4,5 +4,5 @@ package jp.co.soramitsu.core.model
 
 class RuntimeConfiguration(
     val genesisHash: String,
-    val addressByte: Byte,
+    val addressByte: Short,
 )
